rename vertex_helpers.py file
